Rudimentary event handling system

+++ world/generators/chunk/ChunkGenerator.java
@@ -2,8 +2,8 @@ package world.generators.chunk;
 
 public abstract class ChunkGenerator {
 
-    public abstract int[][] generateBase(int size);
-    public abstract int[][] generateObjects(int size);
+    public abstract byte[][] generateBase(int size);
+    public abstract byte[][] generateObjects(int size);
 
     public static ChunkGenerator get(ChunkType type) {
         if (type == ChunkType.OPEN_FIELD) return new OpenFieldChunkGenerator();
